dont delete certain stuff

+++ service-scripts/App-RNASeq.pl
@@ -204,14 +204,16 @@ sub process_rnaseq {
     my $time2 = `date`;
     my $outdir = "$tmpdir";
     
-    # Remove genome directory
-    my $ref_id   = $params->{reference_genome_id};
-    my $ref_dir  = "$tmpdir/$ref_id";
-    rmtree([ "$ref_dir" ]);
+    if ($called_localize_params) {
+        # Remove genome directory
+        my $ref_id   = $params->{reference_genome_id};
+        my $ref_dir  = "$tmpdir/$ref_id";
+        rmtree([ "$ref_dir" ]);
 
-    # Remove TPMCalculator directory
-    my $tpm_dir = "$tmpdir/TPMCalculator";
-    rmtree([ "$tpm_dir" ]);
+        # Remove TPMCalculator directory
+        my $tpm_dir = "$tmpdir/TPMCalculator";
+        rmtree([ "$tpm_dir" ]);
+    }
 
     save_output_files($app,$outdir);
     write_output("Start: $time1"."End:   $time2", "$tmpdir/DONE");
